fix(backend): resolve university DB casting error and implement GAPs 2-4
- fix(db): change Student GPA & EctsGPA fields in Student.cs to double? and remove explicit decimal column type overrides in configuration to match FLOAT type in University DB schema (resolves InvalidCastException)
- fix(repo): include User details when querying Employee records to fix empty supervisor/teacher lists
- feat(gap-2): notify all department students when setting stage periods if SpecialityId is null
- feat(gap-3): add notification to supervisor (direction creator) after direction review (Approve/Reject/Revision)
- fix(gap-4): make review comment optional for Reject decision, requiring it only for RequireRevision
- fix(persistence): call SaveChangesAsync in SetStagesPeriods and ReviewDirection commands to correctly persist dispatched notifications

+++ src/AWM.Service.Domain/University/Student.cs
@@ -12,8 +12,8 @@ public class Student : Entity<int>
     public int? StatusId { get; private set; }
     public int? CategoryId { get; private set; }
     public int Year { get; private set; }
-    public decimal? GPA { get; private set; }
-    public decimal? EctsGPA { get; private set; }
+    public double? GPA { get; private set; }
+    public double? EctsGPA { get; private set; }
     public int? EducationTypeId { get; private set; }
     public int? GrantTypeId { get; private set; }
     public int? AdvisorId { get; private set; }
